+++ nvtabular/columns/schema.py
@@ -32,10 +32,11 @@ from nvtabular.tags import Tags  # noqa
 
 
 def register_extra_metadata(column_schema, feature):
+    filtered_properties = {k: v for k, v in column_schema.properties.items() if k != "domain"}
     msg_struct = Struct()
     # must pack message into "Any" type
     any_pack = Any()
-    any_pack.Pack(json_format.ParseDict(column_schema.properties, msg_struct))
+    any_pack.Pack(json_format.ParseDict(filtered_properties, msg_struct))
     # extra_metadata only takes type "Any" messages
     feature.annotation.extra_metadata.add().CopyFrom(any_pack)
     return feature
@@ -44,10 +45,9 @@ def register_extra_metadata(column_schema, feature):
 def register_list(column_schema, feature):
     if str(column_schema._is_list):
         min_length, max_length = None, None
-        if "min_length" in column_schema.properties:
-            min_length = column_schema.properties["min_length"]
-        if "max_length" in column_schema.properties:
-            max_length = column_schema.properties["max_length"]
+        if "value_count" in column_schema.properties:
+            min_length = column_schema.properties["value_count"]["min"]
+            max_length = column_schema.properties["value_count"]["max"]
         if min_length and max_length and min_length == max_length:
             shape = schema_pb2.FixedShape()
             dim = shape.dim.add()
@@ -62,14 +62,12 @@ def register_list(column_schema, feature):
 
 
 def set_protobuf_float(column_schema, feature):
-    col_min, col_max = 0, 0
-    if "domain" in column_schema.properties:
-        col_min, col_max = column_schema.properties.pop("domain")
+    domain = column_schema.properties.get("domain", {})
     feature.float_domain.CopyFrom(
         schema_pb2.FloatDomain(
             name=column_schema.name,
-            min=col_min,
-            max=col_max,
+            min=domain.get("min", None),
+            max=domain.get("max", None),
         )
     )
     feature.type = schema_pb2.FeatureType.FLOAT
@@ -77,15 +75,16 @@ def set_protobuf_float(column_schema, feature):
 
 
 def set_protobuf_int(column_schema, feature):
-    col_min, col_max = 0, 0
-    if "domain" in column_schema.properties:
-        col_min, col_max = column_schema.properties.pop("domain")
+    domain = column_schema.properties.get("domain", {})
     feature.int_domain.CopyFrom(
         schema_pb2.IntDomain(
             name=column_schema.name,
-            min=col_min,
-            max=col_max,
-            is_categorical="categorical" in column_schema.tags,
+            min=domain.get("min", None),
+            max=domain.get("max", None),
+            is_categorical=(
+                Tags.CATEGORICAL in column_schema.tags
+                or Tags.CATEGORICAL.value in column_schema.tags
+            ),
         )
     )
     feature.type = schema_pb2.FeatureType.INT
@@ -102,8 +101,13 @@ def register_dtype(column_schema, feature):
             string_name = numpy.core._dtype._kind_name(column_schema.dtype)
         elif hasattr(column_schema.dtype, "item"):
             string_name = type(column_schema.dtype(1).item()).__name__
-        else:
+        elif isinstance(column_schema.dtype, str):
+            string_name = column_schema.dtype
+        elif hasattr(column_schema.dtype, "__name__"):
             string_name = column_schema.dtype.__name__
+        else:
+            raise TypeError(f"unsupported dtype for column schema: {column_schema.dtype}")
+
         if string_name in proto_dict:
             feature = proto_dict[string_name](column_schema, feature)
     return feature
@@ -120,14 +124,14 @@ proto_dict = {
 def create_protobuf_feature(column_schema):
     feature = schema_pb2.Feature()
     feature.name = column_schema.name
-    if column_schema.dtype:
-        feature = register_dtype(column_schema, feature)
+    feature = register_dtype(column_schema, feature)
     annotation = feature.annotation
     annotation.tag.extend(
         [tag.value if hasattr(tag, "value") else tag for tag in column_schema.tags]
     )
     # can be instantiated with no values
     # if  so, unnecessary to dump
+    # import pdb; pdb.set_trace()
     if len(column_schema.properties) > 0:
         feature = register_extra_metadata(column_schema, feature)
     return feature
@@ -142,6 +146,10 @@ class ColumnSchema:
     properties: Optional[Dict[str, any]] = field(default_factory=dict)
     dtype: Optional[object] = None
     _is_list: bool = False
+
+    def __post_init__(self):
+        tags = _normalize_tags(self.tags or [])
+        object.__setattr__(self, "tags", tags)
 
     def __str__(self) -> str:
         return self.name
@@ -269,7 +277,6 @@ class Schema:
             dtype = None
             properties = {}
             tags = list(feat.annotation.tag) or []
-            tags = [Tags[tag.upper()] if tag in Tags._value2member_map_ else tag for tag in tags]
             # only one item should ever be in extra_metadata
             if len(feat.annotation.extra_metadata) > 1:
                 raise ValueError(
@@ -288,7 +295,7 @@ class Schema:
                 domain_values = getattr(feat, field_name)
                 # if zero no values were passed
                 if domain_values.max > 0:
-                    properties["domain"] = domain_values.min, domain_values.max
+                    properties["domain"] = {"min": domain_values.min, "max": domain_values.max}
                 if feat.type:
                     if feat.type == 2:
                         dtype = numpy.int
@@ -357,3 +364,7 @@ class Schema:
                 result.column_schemas.pop(key, None)
 
         return result
+
+
+def _normalize_tags(tags):
+    return [Tags[tag.upper()] if tag in Tags._value2member_map_ else tag for tag in tags]